fix: status config tests
* fix: status config tests
fix: deleted old connection test file
refactor: changed method from Gateway interface

* chore: removed some outcommented tests
