+++ gateway/listener_test.go
@@ -85,19 +85,6 @@ func samePK(expected, received protocol.Packet) bool {
 	return sameID && sameData
 }
 
-// type testGateway struct {
-// 	handleCount               int
-// 	returnErrorAfterCallCount int
-// }
-
-// func (gw *testGateway) HandleConnection(conn connection.HSConnection) error {
-// 	gw.handleCount++
-// 	if gw.handleCount >= gw.returnErrorAfterCallCount {
-// 		return ErrStopListenerError
-// 	}
-// 	return nil
-// }
-
 // Actual test methods
 
 // OuterListener Tests
@@ -207,129 +194,3 @@ func TestBasicListener(t *testing.T) {
 		})
 	}
 }
-
-// type parralelServer struct {
-// 	connWg        *sync.WaitGroup
-// 	testWg        *sync.WaitGroup
-// 	receivedConns []connection.HSConnection
-
-// 	blockLogin  bool
-// 	blockStatus bool
-// }
-
-// func (s *parralelServer) Login(conn connection.LoginConnection) error {
-// 	if s.blockLogin {
-// 		s.connWg.Wait()
-// 	}
-// 	s.receivedConns = append(s.receivedConns, conn)
-// 	if !s.blockLogin {
-// 		s.connWg.Done()
-// 	}
-// 	s.testWg.Done()
-// 	return nil
-// }
-// func (s *parralelServer) Status(conn connection.StatusConnection) protocol.Packet {
-// 	if s.blockStatus {
-// 		s.connWg.Wait()
-// 	}
-// 	s.receivedConns = append(s.receivedConns, conn)
-// 	if !s.blockStatus {
-// 		s.connWg.Done()
-// 	}
-// 	s.testWg.Done()
-// 	return protocol.Packet{}
-// }
-
-// type parralelOuterListener struct {
-// 	conns     []connection.Connection
-// 	connCount int
-
-// 	wg                  *sync.WaitGroup
-// 	doneWhenNoMoreConns bool
-// }
-
-// func (s *parralelOuterListener) Start() error {
-// 	return nil
-// }
-
-// func (s *parralelOuterListener) Accept() connection.Connection {
-// 	if s.connCount == len(s.conns) {
-// 		if s.doneWhenNoMoreConns {
-// 			s.wg.Done()
-// 		}
-// 		// To block the thread while kinda acting like a real listener
-// 		//  aka not returning a error bc there are not more connections
-// 		time.Sleep(10 * time.Second)
-// 	}
-// 	tempConn := s.conns[s.connCount]
-// 	s.connCount++
-// 	return tempConn
-// }
-
-// I want to leave it open to make it possible in either Listener or the gateway
-//  to process multiple connection at the same time thats why I dont wanna mock
-//  neither of them in here
-// func TestParallelListening(t *testing.T) {
-
-// 	// Need changes, this cant test or it also support processing multiple
-// 	//  of the same connections types
-// 	connWg := sync.WaitGroup{}
-// 	connWg.Add(1)
-
-// 	testWg := sync.WaitGroup{}
-
-// 	loginPk := login.ServerLoginStart{}.Marshal()
-// 	hs1 := handshaking.ServerBoundHandshake{
-// 		ServerAddress: "infrared-1",
-// 		NextState:     1,
-// 	}
-// 	blockStatus := true
-// 	hs2 := handshaking.ServerBoundHandshake{
-// 		ServerAddress: "infrared-2",
-// 		NextState:     2,
-// 	}
-// 	blockLogin := false
-
-// 	inConn1 := &testInConn{hs: hs1, loginPK: loginPk}
-// 	inConn2 := &testInConn{hs: hs2, loginPK: loginPk}
-// 	inConns := []connection.Connection{inConn1, inConn2}
-// 	testWg.Add(len(inConns))
-
-// 	store := &gateway.SingleServerStore{}
-// 	server := &parralelServer{connWg: &connWg, testWg: &testWg, blockLogin: blockLogin, blockStatus: blockStatus}
-// 	store.Server = server
-// 	tGateway := gateway.CreateBasicGatewayWithStore(store, nil)
-
-// 	outerListener := &parralelOuterListener{conns: inConns}
-// 	listener := &gateway.BasicListener{Gw: &tGateway, OutListener: outerListener}
-
-// 	go func() {
-// 		listener.Listen()
-// 	}()
-// 	channel := make(chan struct{})
-// 	go func(wg *sync.WaitGroup) {
-// 		wg.Wait()
-// 		channel <- struct{}{}
-// 	}(&testWg)
-
-// 	select {
-// 	case <-channel:
-// 		t.Log("Tasked finished before timeout")
-// 	case <-time.After(defaultChanTimeout):
-// 		t.Log("Tasked timed out")
-// 		t.FailNow() // Dont check other code it didnt finish anyway
-// 	}
-
-// 	if len(server.receivedConns) != len(inConns) {
-// 		t.Logf("received %d connection(s) but expected %d", len(server.receivedConns), len(inConns))
-// 		t.FailNow()
-// 	}
-
-// 	if server.receivedConns[0] != inConn2 {
-// 		t.Error("expected conn2 to be received as first but didnt happen")
-// 	}
-// 	if server.receivedConns[1] != inConn1 {
-// 		t.Error("didnt receive conn1 as second connection --which was expected it to be--")
-// 	}
-
-// }
